Module PyreAst.Concrete

This module provides types for a concrete abstract syntax tree of Python, for downstream clients who perfer conventional ADT over the tagless-final approach.

The structure of the syntax tree is kept in sync with the structure of the tagless-final APIs. Consult documentation of TaglessFinal for meanings of the various syntax constructs.

Record/variant definitions are intentionally made private within this module. To construct those records/variants, use the corresponding make_t constructor functions.

module Position : sig ... end

See TaglessFinal.Position.

module Location : sig ... end

See TaglessFinal.Location.

module Identifier : sig ... end

See TaglessFinal.Identifier.

module Constant : sig ... end

See TaglessFinal.Constant.

module ExpressionContext : sig ... end

See TaglessFinal.ExpressionContext.

module BooleanOperator : sig ... end

See TaglessFinal.BooleanOperator.

module BinaryOperator : sig ... end

See TaglessFinal.BinaryOperator.

module UnaryOperator : sig ... end

See TaglessFinal.UnaryOperator.

module ComparisonOperator : sig ... end

See TaglessFinal.ComparisonOperator.

module ImportAlias : sig ... end

See TaglessFinal.ImportAlias.

module Comprehension : sig ... end

See TaglessFinal.Comprehension.

module Keyword : sig ... end

See TaglessFinal.Keyword.

module Argument : sig ... end

See TaglessFinal.Argument.

module Arguments : sig ... end

See TaglessFinal.Arguments.

module Expression : sig ... end

See TaglessFinal.Expression.

module WithItem : sig ... end

See TaglessFinal.WithItem.

module ExceptionHandler : sig ... end

See TaglessFinal.ExceptionHandler.

module MatchCase : sig ... end

See TaglessFinal.MatchCase.

module Pattern : sig ... end

See TaglessFinal.Pattern.

module Statement : sig ... end

See TaglessFinal.Statement.

module TypeIgnore : sig ... end

See TaglessFinal.TypeIgnore.

module Module : sig ... end

See TaglessFinal.Module.

module FunctionType : sig ... end

See TaglessFinal.FunctionType.

val make_tagless_final : unit -> ( Argument.t, Arguments.t, BinaryOperator.t, BooleanOperator.t, ComparisonOperator.t, Comprehension.t, Constant.t, ExceptionHandler.t, Expression.t, ExpressionContext.t, FunctionType.t, Identifier.t, ImportAlias.t, Keyword.t, Location.t, MatchCase.t, Module.t, Pattern.t, Position.t, Statement.t, TypeIgnore.t, UnaryOperator.t, WithItem.t ) TaglessFinal.t

Return a value of TaglessFinal.t whose specification is to construct a concrete AST as defined in this module.