Module Parser.Context

This module contains a type that abstracts away the details of global states required to set up the parser.

type t

An opaque type representing the global state of the parser. Obtaining a value of this type is a pre-requisite for invoking all other parsing-related APIs.

A value of this type can be obtained via Parser.with_context.