Module TaglessFinal.Comprehension

This module provides a type that represents Python comprehension structures.

Example: ... async for x, y in z if x > 0 if y < 0

In this case, target will be (x, y), iter will be z, ifs will be a list [x>0, y<0], and is_async will be true.

type ('expr, 'comprehension) t = target:'expr -> iter:'expr -> ifs:'expr list -> is_async:bool -> 'comprehension