Module TaglessFinal.ExceptionHandler

This module provides a type that represents a Python exception handler.

Example: try ... except Foo as e: pass. Here type_ will be set to Foo, name will be set to e, and the pass part will be the body.

type ('expr, 'identifier, 'location, 'stmt, 'except_handler) t = location:'location -> type_:'expr option -> name:'identifier option -> body:'stmt list -> 'except_handler