Module TaglessFinal.ImportAlias

This module provides a type that represents a Python import item.

name is the target of the import, and asname is the (optional) alias of the target.

type ('identifier, 'location, 'alias) t = location:'location -> name:'identifier -> asname:'identifier option -> 'alias