Module TaglessFinal.Keyword

This module provides a type that represents keyword arguments at each callsite.

arg is the name of the keyword argument, and value is the value specified for that argument.

Reason why the type of arg is an option: If a callsite contains **some_kwarg as argument, it will also be treated as a keyword argument, with arg being unset and value being some_kwarg.

type ('expr, 'identifier, 'location, 'keyword) t = location:'location -> arg:'identifier option -> value:'expr -> 'keyword