Module TaglessFinal.MatchCase

This module provides a type that represents a branch of the Python match statement. See PEP 622.

pattern is the pattern of the branch. guard is the guard expression, if there is one. body is the body of the branch.

type ('expr, 'pattern, 'stmt, 'match_case) t = pattern:'pattern -> guard:'expr option -> body:'stmt list -> 'match_case