Module TaglessFinal.Module

This module provides a type that represents a Python module.

A Python module consists of a collection of statements, paired with a collection of type ignore items.

type ('stmt, 'type_ignore, 'module_) t = body:'stmt list -> type_ignores:'type_ignore list -> 'module_