Module TaglessFinal.Position

A position is a pair of integers representing line number and column numbers.

Line numbers start from 1 and column numbers start from 0. In other words, first character in the input string has location of line 1, column 0.

type 'a t = line:int -> column:int -> 'a