Module TaglessFinal.TypeIgnore

This module provides a type that represents a Python type ignore item.

A type ignore item can only appear as comment in the source code, in the form of # type: ignore.

lineno keeps track of their line numbers. And tag keep track of anything else that comes after the ignore comment.

type 'a t = lineno:int -> tag:string -> 'a