Module PyreAst

module TaglessFinal : sig ... end

This module provides tagless-final interfaces for the syntax of Python.

module Concrete : sig ... end

This module provides types for a concrete abstract syntax tree of Python, for downstream clients who perfer conventional ADT over the tagless-final approach.

module Parser : sig ... end

This module contains all parsing APIs, i.e. functions that transfrom plain strings into syntactical constructs. These APIs will always hand back a Result.t where if the parsing fails, a Parser.Error.t gets returned.