Module TaglessFinal.ExpressionContext

This module provides a type that represents Python expression contexts.

Expression contexts are mainly used to syntactically differentiate among expressions that will be read at runtime, those that will be written at runtime, and those that will be deled at runtime.

As of Python 3.10, there are 6 kinds of expressions with an explicit context field:

Expression.t.name

Expression.t.attribute

Expression.t.subscript

Expression.t.list

Expression.t.tuple

Expression.t.starred

These expressions are allowed to appear at the following "target" positions:

The targets field in Statement.t.assign.

The target field in Statement.t.aug_assign and Statement.t.ann_assign (except for list, tuple, and starred expression).

The targets field in Statement.t.delete (except for starred expression).

The targets field in Statement.t.for_ (except for starred expression).

The targets field in Statement.t.with_ (except for starred expression).

type 'a t = private {

load : 'a;

store : 'a;

del : 'a;

}

val make : load:'a -> store:'a -> del:'a -> unit -> 'a t

Constructor of t.