Module TaglessFinal.Identifier

This module provide a type that represents Python identifiers.

Due to technical limitation, currently this library does not support parsing non-ascii identifiers as specified in PEP 3131.

type 'a t = string -> 'a