Module TaglessFinal.Location

A location is a pair of Position.t representing a range for a given token.

Ranges are inclusive in their start position and exclusive in their end position. For example, identifier foo at the begining of an input string has a location from (line 1, column 0) to (line 1, column 3).

type ('position, 'location) t = start:'position -> stop:'position -> 'location